fix round function
